Adds two implementations of Jaiqu using Burr
jaiqu_burr_straight_port takes the code and
maps it directly to burr.

jaiqu_burr_granular takes the straight
port and breaks it up into more actions.

Note: I was hoping to use the mermaid
diagram and do something reflecting that
but the logic in the mermaid diagram
doesn't really map to the code and
thus isn't a good thing to model -- but
the more granualr version does kind of
look like it.